# backend/docs/swagger.yaml
basePath: /
definitions:
  res.ErrorConstant:
    properties:
      code:
        type: integer
      data: {}
      error_fields:
        items:
          $ref: '#/definitions/res.Message'
        type: array
      message:
        type: string
      status:
        type: boolean
    type: object
  res.Message:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
...
  schemas.CreateAddress:
    properties:
      address_label:
        type: string
      city:
        type: string
      district:
        type: string
      full_address:
        type: string
      is_default:
        type: boolean
      latitude:
        type: number
      longitude:
        type: number
      phone_number:
        type: string
      postal_code:
        type: string
      province:
        type: string
      receiver_name:
        type: string
      village:
        type: string
    required:
    - city
    - district
    - full_address
    - phone_number
    - postal_code
    - province
    - receiver_name
    type: object
...
  schemas.UpdateAddress:
    properties:
      address_label:
        type: string
      city:
        type: string
      district:
        type: string
      full_address:
        type: string
      is_default:
        type: boolean
      latitude:
        type: number
      longitude:
        type: number
      phone_number:
        type: string
      postal_code:
        type: string
      province:
        type: string
      receiver_name:
        type: string
      village:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: E-Commerce NW SBO Backend API
  title: E-Commerce NW API
  version: "1.0"
paths:
  /addresses/{id}/addresses:
    get:
      consumes:
      - application/json
      description: Get all addresses for a specific user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.ErrorConstant'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.ErrorConstant'
      summary: GET User Addresses
      tags:
      - Addresses
    post:
      consumes:
      - application/json
      description: Create a new address for a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Address data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.CreateAddress'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/res.ErrorConstant'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.ErrorConstant'
      summary: POST User Address
      tags:
      - Addresses
  /addresses/{id}/addresses/{address_id}:
    delete:
      consumes:
      - application/json
      description: Delete an address for a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Address ID
        in: path
        name: address_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.ErrorConstant'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.ErrorConstant'
      summary: DELETE User Address
      tags:
      - Addresses
    get:
      consumes:
      - application/json
      description: Get a specific address by ID for a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Address ID
        in: path
        name: address_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.ErrorConstant'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.ErrorConstant'
      summary: GET User Address by ID
      tags:
      - Addresses
    put:
      consumes:
      - application/json
      description: Update an existing address for a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Address ID
        in: path
        name: address_id
        required: true
        type: string
      - description: Address data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.UpdateAddress'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.ErrorConstant'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.ErrorConstant'
      summary: PUT User Address
      tags:
      - Addresses
  /addresses/{id}/addresses/{address_id}/default:
    put:
      consumes:
      - application/json
      description: Set an address as the default address for a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Address ID
        in: path
        name: address_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.ErrorConstant'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.ErrorConstant'
      summary: PUT Set Default Address
      tags:
      - Addresses
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"